Use user's locale to determine which version of the Briefing is served.
