--- functions/src/snapshotterFunction/models.ts
@@ -1,7 +1,0 @@
-enum Locale {
-  AU = 'en-AU',
-  GB = 'en-GB',
-  US = 'en-US',
-}
-
-export { Locale };
