+++ functions/src/snapshotterFunction/snapshotterFunction.ts
@@ -3,7 +3,7 @@ import * as fs from 'fs-extra';
 import { Storage } from '@google-cloud/storage';
 import { region, config } from 'firebase-functions';
 import { Response } from 'express';
-import { Locale } from './models';
+import { Locale, getLocale } from '../localeUtils';
 
 const appConfig = config();
 
@@ -28,20 +28,6 @@ const generateBriefing = region('europe-west1').https.onRequest(
       });
   }
 );
-
-const getLocale = (locale: any) => {
-  const supportedLocales: { [key: string]: Locale } = {
-    'en-AU': Locale.AU,
-    'en-CA': Locale.US,
-    'en-GB': Locale.GB,
-    'en-IN': Locale.GB,
-    'en-SG': Locale.GB,
-    'en-US': Locale.US,
-  };
-
-  const l: Locale = supportedLocales[locale];
-  return typeof l === 'undefined' ? Locale.GB : l;
-};
 
 const buildURL = (locale: Locale) => {
   switch (locale) {
